pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "slack-emoji-tailor"
version = "0.1.0"
description = "Slack emoji conversion web service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.0",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
slack-emoji-tailor = "app.main:run_server"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
